requirements.txt
streamlit
langchain
langchain-nvidia-ai-endpoints
pypdf
networkx
pyvis
python-dotenv